+++ packages/column-views/src/maps/_shared/index.ts
@@ -1,2 +1,0 @@
-export * from "./inset-map";
-export * from "./styles";