+++ packages/column-views/src/maps/column-navigation/mapbox/index.ts
@@ -1,21 +1,17 @@
-import { useMapStyleOperator } from "@macrostrat/mapbox-react";
+import { useMapRef, useMapStyleOperator } from "@macrostrat/mapbox-react";
 import h from "@macrostrat/hyper";
-import { FeatureCollection } from "geojson";
-import { ReactNode, useMemo, useRef } from "react";
-import { buildColumnsStyle, InsetMapProps } from "../../_shared";
-import { setGeoJSON } from "@macrostrat/mapbox-utils";
+import { ReactNode, useCallback, useEffect, useMemo, useRef } from "react";
 
 import {
   ColumnNavigationProvider,
   NavigationProviderProps,
   useColumnNavigationStore,
 } from "./state";
-import { InsetMap } from "../../_shared";
+import { InsetMap, type InsetMapProps } from "../../inset-map";
 import { geoCentroid } from "d3-geo";
-import {
-  ColumnKeyboardNavigation,
-  buildKeyboardNavigationStyle,
-} from "./keyboard-navigation";
+import { BaseColumnsLayer, SelectedColumnOverlay } from "../../layers";
+import { ColumnKeyboardNavigation } from "./keyboard-navigation";
+import type { ColumnFeature } from "../../layers";
 
 export interface ColumnNavigationMapProps extends InsetMapProps {
   padding?: number;
@@ -27,43 +23,11 @@ export interface ColumnNavigationMapProps extends InsetMapProps {
   onHoverColumn?: (colID: number | null) => void;
   showTriangulation?: boolean;
   columnColor?: string;
-  triangulationColor?: string;
-  keyboardNavigation?: boolean;
 }
 
 export function ColumnNavigationMap(
   props: ColumnNavigationMapProps & NavigationProviderProps,
 ) {
-  const {
-    columns,
-    columnIDs,
-    onSelectColumn,
-    onHoverColumn,
-    selectedColumn,
-    hoveredColumn,
-    projectID,
-    children,
-    inProcess,
-    ...rest
-  } = props;
-
-  return h(
-    ColumnNavigationProvider,
-    {
-      columns,
-      columnIDs,
-      onSelectColumn,
-      onHoverColumn,
-      selectedColumn,
-      hoveredColumn,
-      projectID,
-      inProcess,
-    },
-    h(_ColumnNavigationMap, props),
-  );
-}
-
-function _ColumnNavigationMap(props: ColumnNavigationMapProps) {
   const {
     padding = 50,
     children,
@@ -71,16 +35,16 @@ function _ColumnNavigationMap(props: ColumnNavigationMapProps) {
     showTriangulation,
     triangulationColor,
     keyboardNavigation = true,
+    columns,
+    columnIDs,
+    onSelectColumn,
+    onHoverColumn,
+    selectedColumn,
+    hoveredColumn,
+    projectID,
+    inProcess,
     ...rest
   } = props;
-
-  const overlayStyles = useMemo(() => {
-    let styles: any[] = [buildColumnsStyle(columnColor)];
-    if (showTriangulation) {
-      styles.push(buildKeyboardNavigationStyle(triangulationColor));
-    }
-    return styles;
-  }, [columnColor, showTriangulation, triangulationColor]);
 
   return h(
     InsetMap,
@@ -88,136 +52,225 @@ function _ColumnNavigationMap(props: ColumnNavigationMapProps) {
       ...rest,
       boxZoom: false,
       dragRotate: false,
-      overlayStyles,
     },
     [
-      h(ColumnsLayer),
+      h(ColumnsNavigationLayer, {
+        color: columnColor,
+        columns,
+        columnIDs,
+        onSelectColumn,
+        onHoverColumn,
+        selectedColumn,
+        hoveredColumn,
+        projectID,
+        inProcess,
+        keyboardNavigation,
+        showTriangulation,
+        triangulationColor,
+      }),
       children,
-      h.if(keyboardNavigation)(ColumnKeyboardNavigation, { showTriangulation }),
     ],
   );
 }
 
-function ColumnsLayer({ enabled = true }) {
+interface ColumnLayerProps {
+  enabled?: boolean;
+  color?: string;
+}
+
+interface ColumnKeyboardNavigationProps {
+  triangulationColor?: string;
+  keyboardNavigation?: boolean;
+  showTriangulation?: boolean;
+}
+
+export function ColumnsNavigationLayer(
+  props: NavigationProviderProps &
+    ColumnLayerProps &
+    ColumnKeyboardNavigationProps,
+) {
+  const {
+    enabled,
+    color,
+    keyboardNavigation,
+    showTriangulation,
+    triangulationColor,
+    ...providerProps
+  } = props;
+
+  return h(ColumnNavigationProvider, providerProps, [
+    h(_ColumnsNavigationLayer, { enabled, color }),
+    h.if(keyboardNavigation)(ColumnKeyboardNavigation, {
+      showTriangulation,
+      triangulationColor,
+    }),
+  ]);
+}
+
+const columnLayers = ["columns-points", "columns-fill"];
+
+function _ColumnsNavigationLayer({ enabled = true, color }) {
   const columns = useColumnNavigationStore((state) => state.columns);
-  const selectedColumn = useColumnNavigationStore(
+  const selectedColumnID = useColumnNavigationStore(
     (state) => state.selectedColumn,
-  );
-  const hoveredColumn = useColumnNavigationStore(
-    (state) => state.hoveredColumn,
   );
   const selectColumn = useColumnNavigationStore((state) => state.selectColumn);
   const setHoveredColumn = useColumnNavigationStore(
     (state) => state.setHoveredColumn,
   );
 
+  const onHover = useCallback(
+    (event) => setHoveredColumn(event.features?.[0]?.id),
+    [setHoveredColumn],
+  );
+
+  /** Convert features into column IDs to pass to the selection manager */
+  const onSelectColumn = useCallback(
+    (column) => selectColumn(column.id),
+    [selectColumn],
+  );
+
+  const selectedColumnFeature = useMemo(() => {
+    if (columns == null || selectedColumnID == null) return null;
+    return columns.find((d) => d.id == selectedColumnID);
+  }, [selectedColumnID, columns]);
+
+  return h([
+    h(BaseColumnsLayer, { enabled, color, columns }),
+    h(ColumnSelectionManager, {
+      selectedColumn: selectedColumnFeature,
+      onSelectColumn,
+    }),
+    h(ColumnHoverInteraction, { onHover }),
+  ]);
+}
+
+export function ColumnSelectionManager({
+  selectedColumn,
+  onSelectColumn,
+  centerOnSelect = true,
+}: {
+  selectedColumn: ColumnFeature;
+  onSelectColumn: (column: ColumnFeature) => void;
+  centerOnSelect?: boolean;
+}) {
+  const onClick = useCallback(
+    (event) => {
+      if (event.features.length == 0) return;
+      const selectedColumn = event.features[0];
+      onSelectColumn(selectedColumn);
+    },
+    [onSelectColumn],
+  );
+
+  return h([
+    h.if(centerOnSelect)(SelectedColumnCenterer, {
+      selectedColumn,
+    }),
+    h(SelectedColumnOverlay, { selectedColumn }),
+    h(ClickInteraction, {
+      onClick,
+      layers: columnLayers,
+    }),
+  ]);
+}
+
+export function ColumnHoverInteraction({ onHover }) {
+  return h(HoverInteraction, { onHover, layers: columnLayers });
+}
+
+function ClickInteraction({ onClick, layers }) {
   useMapStyleOperator(
     (map) => {
-      if (columns == null) {
-        return;
-      }
-      const data: FeatureCollection = {
-        type: "FeatureCollection",
-        features: columns,
+      // Setup hover styles on map
+      map.on("click", layers, onClick);
+      return () => {
+        map.off("click", layers, onClick);
       };
+    },
+    [layers, onClick],
+  );
+  return null;
+}
 
-      setGeoJSON(map, "columns", data);
+function HoverInteraction({ onHover, layers }) {
+  const hoveredFeaturesRef = useRef([]);
+  const updateFeatureState = useCallback((map, features) => {
+    // Reset hover state for previously hovered features
+    for (const f of hoveredFeaturesRef.current) {
+      map.setFeatureState(
+        { source: f.source, id: f.id, sourceLayer: f.sourceLayer },
+        { hover: false },
+      );
+    }
+    // Set hover state for new features
+    for (const f of features) {
+      map.setFeatureState(
+        { source: f.source, id: f.id, sourceLayer: f.sourceLayer },
+        { hover: true },
+      );
+    }
+    hoveredFeaturesRef.current = features;
+  }, []);
 
+  useMapStyleOperator(
+    (map) => {
       const mouseMove = (event) => {
         if (event.features.length == 0) return;
-        const hoveredFeature = event.features[0];
-        setHoveredColumn(hoveredFeature.id);
+        onHover?.(event);
+        updateFeatureState(map, event.features);
         map.getCanvas().style.cursor = "pointer";
       };
-      const mouseLeave = () => {
-        setHoveredColumn(null);
+      const mouseLeave = (event) => {
+        onHover?.(event);
+        updateFeatureState(map, event.features ?? []);
         map.getCanvas().style.cursor = "";
       };
 
-      const clickHandler = (event) => {
-        if (event.features.length == 0) return;
-        const selectedColumn = event.features[0];
-        const id = selectedColumn.id;
-        selectColumn(id);
-      };
-
       // Setup hover styles on map
-      const layers = ["columns-points", "columns-fill"];
       map.on("mousemove", layers, mouseMove);
       map.on("mouseleave", layers, mouseLeave);
-      map.on("click", layers, clickHandler);
 
       return () => {
         map.off("mouseenter", layers, mouseMove);
         map.off("mouseleave", layers, mouseLeave);
-        map.off("click", layers, clickHandler);
       };
     },
-    [columns, enabled],
+    [layers, onHover],
   );
+  return null;
+}
 
-  const hoveredColumnRef = useRef(null);
-  useMapStyleOperator(
-    (map) => {
-      const prevHoveredColumn = hoveredColumnRef.current;
-      if (hoveredColumn == prevHoveredColumn) return;
-      if (prevHoveredColumn != null) {
-        // Deselect previous column
-        map.setFeatureState(
-          { source: "columns", id: prevHoveredColumn },
-          { hover: false },
-        );
-      }
-      hoveredColumnRef.current = hoveredColumn;
-      // Select the current column
-      map.setFeatureState(
-        { source: "columns", id: hoveredColumn },
-        { hover: true },
-      );
-    },
-    [hoveredColumn, columns],
-  );
+function SelectedColumnCenterer({
+  selectedColumn,
+  easeDuration = 500,
+}: {
+  selectedColumn: ColumnFeature;
+  easeDuration?: number;
+}) {
+  const columnCenter = useMemo(() => {
+    const columnGeometry = selectedColumn?.geometry;
+    if (columnGeometry == null) return null;
+    return geoCentroid(columnGeometry);
+  }, [selectedColumn]);
 
-  /** Set feature state for selected columns */
-  const selectedColumnRef = useRef(null);
+  const mapRef = useMapRef();
+  const map = mapRef.current;
   const initialRenderRef = useRef(true);
-  useMapStyleOperator(
-    (map) => {
-      if (columns == null) return;
-      const prevSelectedColumn = selectedColumnRef.current;
-      if (selectedColumn == prevSelectedColumn) return;
-      if (prevSelectedColumn != null) {
-        // Deselect previous column
-        map.setFeatureState(
-          { source: "columns", id: prevSelectedColumn },
-          { selected: false },
-        );
-      }
-
-      selectedColumnRef.current = selectedColumn;
-
-      // Select the current column
-      map.setFeatureState(
-        { source: "columns", id: selectedColumn },
-        { selected: true },
-      );
-
-      // Center the selected column
-      const columnGeometry = columns.find(
-        (d) => d.id == selectedColumn,
-      )?.geometry;
-      if (columnGeometry == null) return;
-      const center = geoCentroid(columnGeometry);
-
-      const isInitialRender = initialRenderRef.current;
-
-      map.easeTo({ center }, { duration: isInitialRender ? 0 : 500 });
-      if (isInitialRender) {
-        initialRenderRef.current = false;
-      }
-    },
-    [selectedColumn, columns],
-  );
+  // Center the map on the selected column when it changes
+  useEffect(() => {
+    if (map == null) return;
+    const isInitialRender = initialRenderRef.current;
+    if (isInitialRender) initialRenderRef.current = false;
+    // If this is the first render of the map, jump to the location. Otherwise, ease to it.
+    if (columnCenter == null) return;
+    if (isInitialRender) {
+      map.setCenter(columnCenter);
+      initialRenderRef.current = false;
+    } else {
+      map.easeTo({ center: columnCenter }, { duration: easeDuration });
+    }
+  }, [mapRef.current, columnCenter]);
 
   return null;
 }