+++ packages/column-views/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file. The format
 is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.3.0] - 2026-02-13
+
+- Greatly streamline handling of `InsetMap`, column and unit overlay layers
+- Improve design of age range tags
+
 ## [3.2.2] - 2026-02-09
 
 - Move `AgeField` and `AgeRange` to separate exports