+++ packages/map-interface/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file. The format
 is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.0] - 2026-02-13
+
+- Add internal logic to `MapView` to manage the addition of custom style
+  fragments, based on updates to the `@macrostrat/mapbox-react` package
+- Add new stories
+
 ## [2.1.0] - 2026-02-09
 
 - Remove xDD info panel components